requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.3
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.14.6
chardet==4.0.0
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.1.3
mathparse==0.1.2
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.2.0
Pint==0.18
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pygame==2.0.1
Pygments==2.9.0
PyJWT==1.7.1
pymongo==3.12.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.7.5
python-dotenv==1.0.0
pytz==2021.1
PyYAML==3.13
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
redis==5.0.1
regex==2021.8.28
requests==2.31.0
requests-oauthlib==1.3.0
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
spotipy==2.23.0
SQLAlchemy==1.2.19
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tweepy==4.4.0
twilio==6.60.0
urllib3==2.1.0
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==3.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.17.0
